feat: show unscheduled activites on list page
