+++ src/Activity/ActivityIdea.module.css
@@ -1,0 +1,51 @@
+.activityCard {
+  border: 1px solid var(--gray-9);
+  background-color: var(--color-panel-solid);
+  border-radius: var(--radius-2);
+  overflow: hidden;
+  cursor: var(--cursor-link);
+  display: block;
+  text-align: start;
+
+  &:hover {
+    border-color: var(--accent-7);
+    background-color: var(--accent-2);
+    transform: translateY(-1px);
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  }
+
+  &[draggable="true"] {
+    cursor: grab;
+
+    &:active {
+      cursor: grabbing;
+    }
+  }
+
+  &:focus-visible {
+    outline: 2px solid var(--accent-9);
+    outline-offset: 2px;
+  }
+
+  &.draggingCard {
+    opacity: 0.6;
+    transform: rotate(2deg);
+    cursor: grabbing;
+    box-shadow: 0 0 8px rgba(0, 0, 0, 0.2);
+  }
+}
+
+.activityTitle {
+  line-height: 1.3;
+  color: var(--color-text);
+}
+
+.activityLocation,
+.activityDescription {
+  line-height: 1.4;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+}
