+++ src/Ideas/IdeaSidebar.module.css
@@ -66,54 +66,9 @@
   gap: var(--space-2);
 }
 
-.activityCard {
+.activityItem {
   padding: var(--space-3);
   border: 1px solid var(--color-border);
-  border-radius: var(--radius-2);
-  cursor: pointer;
-  transition: all 0.2s ease;
-  background-color: var(--color-surface);
-
-  &:hover {
-    border-color: var(--accent-7);
-    background-color: var(--accent-2);
-    transform: translateY(-1px);
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  }
-
-  &[draggable="true"] {
-    cursor: grab;
-
-    &:active {
-      cursor: grabbing;
-    }
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--accent-7);
-    outline-offset: 2px;
-  }
-
-  &.draggingCard {
-    opacity: 0.5;
-    transform: rotate(2deg);
-    cursor: grabbing;
-  }
-}
-
-.activityTitle {
-  line-height: 1.3;
-  color: var(--color-text);
-}
-
-.activityLocation,
-.activityDescription {
-  line-height: 1.4;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
 }
 
 .locationIcon,
