+++ src/Ideas/IdeaSidebar.tsx
@@ -1,21 +1,8 @@
-import {
-  CalendarIcon,
-  ClockIcon,
-  InfoCircledIcon,
-  SewingPinIcon,
-} from '@radix-ui/react-icons';
-import {
-  Badge,
-  Box,
-  Card,
-  ContextMenu,
-  Flex,
-  ScrollArea,
-  Text,
-} from '@radix-ui/themes';
+import { CalendarIcon, ClockIcon } from '@radix-ui/react-icons';
+import { Badge, Box, Card, ScrollArea, Text } from '@radix-ui/themes';
 import clsx from 'clsx';
-import { memo, useCallback, useMemo, useState } from 'react';
-import { useActivityDialogHooks } from '../Activity/activityDialogHooks';
+import { memo, useMemo } from 'react';
+import { ActivityIdea } from '../Activity/ActivityIdea';
 import type { TripSliceActivity } from '../Trip/store/types';
 import { TripViewMode } from '../Trip/TripViewMode';
 import s from './IdeaSidebar.module.css';
@@ -24,114 +11,6 @@ interface IdeaSidebarProps {
   activities: TripSliceActivity[];
   userCanEditOrDelete: boolean;
   isVisible: boolean;
-}
-
-interface IdeaActivityCardProps {
-  activity: TripSliceActivity;
-  userCanEditOrDelete: boolean;
-}
-
-function IdeaActivityCard({
-  activity,
-  userCanEditOrDelete,
-}: IdeaActivityCardProps) {
-  const [isDragging, setIsDragging] = useState(false);
-  const {
-    openActivityViewDialog,
-    openActivityDeleteDialog,
-    openActivityEditDialog,
-  } = useActivityDialogHooks(TripViewMode.Timetable, activity.id);
-  const handleDragStart = useCallback(
-    (e: React.DragEvent<HTMLDivElement>) => {
-      setIsDragging(true);
-      // Store the activity data for the drop
-      e.dataTransfer.setData(
-        'text/plain',
-        JSON.stringify({
-          activityId: activity.id,
-          originalTimeStart: null,
-          originalTimeEnd: null,
-          originalDayStart: null,
-        }),
-      );
-
-      // Set the drag image/opacity
-      e.dataTransfer.effectAllowed = 'move';
-      if (e.currentTarget) {
-        e.dataTransfer.setDragImage(e.currentTarget, 20, 20);
-      }
-    },
-    [activity.id],
-  );
-
-  const handleDragEnd = useCallback(() => {
-    setIsDragging(false);
-  }, []);
-
-  const handleKeyDown = useCallback(
-    (e: React.KeyboardEvent) => {
-      if (e.key === 'Enter' || e.key === ' ') {
-        e.preventDefault();
-        openActivityViewDialog();
-      }
-    },
-    [openActivityViewDialog],
-  );
-
-  return (
-    <ContextMenu.Root>
-      <ContextMenu.Trigger>
-        <Card
-          className={clsx(s.activityCard, isDragging && s.draggingCard)}
-          draggable={userCanEditOrDelete}
-          onDragStart={handleDragStart}
-          onDragEnd={handleDragEnd}
-          onClick={openActivityViewDialog}
-          onKeyDown={handleKeyDown}
-          tabIndex={0}
-          aria-label={`Activity: ${activity.title}${activity.location ? `, at ${activity.location}` : ''}`}
-        >
-          <Flex direction="column" gap="1">
-            <Text size="2" weight="medium" className={s.activityTitle}>
-              {activity.title}
-            </Text>
-            {activity.location && (
-              <Flex align="center" gap="1">
-                <SewingPinIcon className={s.locationIcon} />
-                <Text size="1" color="gray" className={s.activityLocation}>
-                  {activity.location}
-                </Text>
-              </Flex>
-            )}
-            {activity.description && (
-              <Flex align="center" gap="1">
-                <InfoCircledIcon className={s.descriptionIcon} />
-                <Text size="1" color="gray" className={s.activityDescription}>
-                  {activity.description}
-                </Text>
-              </Flex>
-            )}
-          </Flex>
-        </Card>
-      </ContextMenu.Trigger>
-      <ContextMenu.Content>
-        <ContextMenu.Item onClick={openActivityViewDialog}>
-          View
-        </ContextMenu.Item>
-        {userCanEditOrDelete && (
-          <>
-            <ContextMenu.Item onClick={openActivityEditDialog}>
-              Edit
-            </ContextMenu.Item>
-            <ContextMenu.Separator />
-            <ContextMenu.Item onClick={openActivityDeleteDialog} color="red">
-              Delete
-            </ContextMenu.Item>
-          </>
-        )}
-      </ContextMenu.Content>
-    </ContextMenu.Root>
-  );
 }
 
 function IdeaSidebarInner({
@@ -168,10 +47,12 @@ function IdeaSidebarInner({
       <ScrollArea className={s.content}>
         <Box className={s.activityList}>
           {ideaActivities.map((activity) => (
-            <IdeaActivityCard
+            <ActivityIdea
               key={activity.id}
               activity={activity}
               userCanEditOrDelete={userCanEditOrDelete}
+              tripViewMode={TripViewMode.Timetable}
+              className={s.activityItem}
             />
           ))}
         </Box>
